Add PlayerRole class and formation file readers to handle player roles
